requirements.txt
# Universal Converter V3.0.0 - Python 3.12+
# ==========================================
# Kurulum: pip install -r requirements.txt -q

# === Temel Bağımlılıklar ===
fastapi
uvicorn
python-multipart
pillow
pandas
openpyxl
PyPDF2
python-docx
lxml

# === NUMPY UYARISI ===
# pdf2docx numpy>=2.0 ile uyumsuz, bu yüzden sınırlıyoruz
numpy<2.0

# === Gelişmiş Belge Dönüştürme ===
pdf2docx
docx2pdf
python-pptx
PyMuPDF
reportlab

# === Ek Özellikler ===
py7zr
pillow-heif
cairosvg

# === Windows Özel (PPTX/DOCX → PDF) ===
comtypes; platform_system == "Windows"
